moving logreg models from a subfold to the main fold
