+++ tests/trainer/baseline/test_baseline.py
@@ -2,7 +2,7 @@ import numpy as np
 import torch
 from torch.utils.data import TensorDataset
 
-from knodle.model.logistic_regression.logistic_regression_model import LogisticRegressionModel
+from knodle.model.logistic_regression_model import LogisticRegressionModel
 from knodle.trainer.baseline.no_denoising import NoDenoisingTrainer
