+++ tutorials/__init__.py
@@ -6,9 +6,6 @@ __author__ = "knodle <knodle@cs.univie.ac.at>"
 __all__ = []
 
 import logging
-from knodle.model.logistic_regression.logistic_regression_model import (
-    LogisticRegressionModel,
-)
 
 logger = logging.getLogger()
 handler = logging.StreamHandler()
